requirements.txt

python-bitcoinrpc==1.0
toml==0.10.1
